update img preview

+++ components/preview-img.tsx
@@ -1,10 +1,6 @@
 import { XIcon } from "lucide-react";
 import { Button } from "./ui/button";
-import {
-  FileWithPreview,
-  formatBytes,
-  useFileUpload,
-} from "@/hooks/use-file-upload";
+import type { FileWithPreview } from "@/hooks/use-file-upload";
 import { useGlobalstate } from "@/context/global-store";
 
 export default function PreviewImg({
@@ -16,25 +12,28 @@ export default function PreviewImg({
   removeFile: (id: string) => void;
   clearFiles: () => void;
 }) {
-  const { setFileExists } = useGlobalstate();
+  const { storedFiles, setStoredFiles, setFileExists } = useGlobalstate();
 
   return (
     <>
-      {/* File list */}
       {files.length > 0 && (
-        <div className="space-y-2 absolute -top-12 isolate z-10 bg-white left-14 ">
+        <div className="space-y-2 absolute -top-12 isolate z-10 bg-white left-14">
           {files.map((file) => {
+            const previewUrl = (file.file as any).base64 || file.preview;
+
+            if (!previewUrl) return null;
+
             return (
               <div
                 key={file.id}
-                className="bg-background flex items-center justify-between gap-2 rounded-lg border  "
+                className="bg-background flex items-center justify-between gap-2 rounded-lg border"
               >
-                <div className="flex items-center gap-3 ">
-                  <div className="bg-accent aspect-square shrink-0 rounded relative ">
+                <div className="flex items-center gap-3">
+                  <div className="bg-accent aspect-square shrink-0 rounded relative">
                     <Button
                       size="icon"
                       variant="ghost"
-                      className="text-muted-foreground/80 hover:text-foreground -me-2 size-8 hover:bg-transparent absolute -top-4 -right-2 rounded-full    "
+                      className="text-muted-foreground/80 hover:text-foreground -me-2 size-8 hover:bg-transparent absolute -top-4 -right-2 rounded-full"
                       onClick={() => {
                         removeFile(file.id);
                         setFileExists(false);
@@ -44,7 +43,7 @@ export default function PreviewImg({
                       <XIcon aria-hidden="true" color="red" />
                     </Button>
                     <img
-                      src={file.preview}
+                      src={previewUrl}
                       alt={file.file.name}
                       className="size-14 rounded-[inherit] object-cover"
                     />
@@ -54,7 +53,6 @@ export default function PreviewImg({
             );
           })}
 
-          {/* Remove all files button */}
           {files.length > 1 && (
             <div>
               <Button size="sm" variant="outline" onClick={clearFiles}>
